move WorkManagerRepository to workmanager package
